[#18] refactor - common 폴더에 constans 모아놓음
